# docs/swagger.yaml
definitions:
  api-gateway_pkg_api_auth_v1.GetUsersInfoResponse:
    properties:
      users:
        additionalProperties:
          $ref: '#/definitions/api-gateway_pkg_api_auth_v1.UserInfo'
        description: Возвращаем map, где ключ - user_id.
        type: object
    type: object
  api-gateway_pkg_api_auth_v1.LoginResponse:
    properties:
      access_token:
        description: JWT-токен
        type: string
    type: object
  api-gateway_pkg_api_auth_v1.RegisterResponse:
    properties:
      access_token:
        description: JWT-токен
        type: string
    type: object
  api-gateway_pkg_api_auth_v1.UserInfo:
    properties:
      id:
        type: string
      login:
        type: string
      name:
        type: string
    type: object
...
  api-gateway_pkg_response.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  internal_transport_handlers_authorization.GetTelegramResponse:
    properties:
      link:
        type: string
    type: object
  internal_transport_handlers_authorization.GetUsersInfoDTO:
    properties:
      user_ids:
        items:
          type: string
        type: array
    type: object
  internal_transport_handlers_authorization.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_transport_handlers_authorization.RegisterDTO:
    properties:
      email:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
  title: Eventify API Gateway
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Принимает email и пароль, возвращает JWT токен
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handlers_authorization.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: вернёт access_token
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_auth_v1.LoginResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "401":
          description: Неверный логин или пароль
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает JWT токен
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handlers_authorization.RegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: вернет access_token
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_auth_v1.RegisterResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "409":
          description: Пользователь с таким email или логином уже существует
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /users/{login}:
    get:
      description: Ищет пользователя по уникальному логину (никнейму) и возвращает
        его публичные данные
      parameters:
      - description: Логин пользователя (без @)
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Публичные данные пользователя
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_auth_v1.UserInfo'
        "400":
          description: Логин не передан
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      summary: Получение публичного профиля пользователя
      tags:
      - users
  /users/info:
    post:
      consumes:
      - application/json
      description: Возвращает мапу с данными пользователей по переданному массиву
        ID
      parameters:
      - description: Массив UUID пользователей
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handlers_authorization.GetUsersInfoDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Map пользователей (ключ - ID)
          schema:
            $ref: '#/definitions/api-gateway_pkg_api_auth_v1.GetUsersInfoResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "404":
          description: Пользователи не найдены
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      summary: Получение информации о группе пользователей
      tags:
      - users
  /users/telegram_link:
    post:
      description: Генерирует уникальную ссылку с токеном для привязки аккаунта Telegram
        к профилю пользователя. Ссылка одноразовая.
      produces:
      - application/json
      responses:
        "200":
          description: Ссылка на бота
          schema:
            $ref: '#/definitions/internal_transport_handlers_authorization.GetTelegramResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api-gateway_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить ссылку на Telegram-бота
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: API для сервиса мероприятий Eventify
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"